requirements.txt
prettytable==2.4.0
py-gitea==0.1.3
PyGithub==1.55
